fix migration

+++ Modules/Doctor/Database/Migrations/2022_07_27_141942_update_drop_doctor_inboxes_column_to_doctor_inboxes_table.php
@@ -4,7 +4,7 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class UpdateDoctorInboxesColumnToDoctorInboxesTable extends Migration
+class UpdateDropDoctorInboxesColumnToDoctorInboxesTable extends Migration
 {
     public function __construct() {
         DB::getDoctrineSchemaManager()->getDatabasePlatform()->registerDoctrineTypeMapping('enum', 'string');
@@ -16,10 +16,13 @@ class UpdateDoctorInboxesColumnToDoctorInboxesTable extends Migration
      */
     public function up()
     {
-        Schema::table('doctor_inboxes', function (Blueprint $table) {
-            $table->dropColumn('inboxes_promotion_status');
-			$table->string('inboxes_category')->after('inboxes_id_reference');
-        });
+        if (Schema::hasColumn('doctor_inboxes', 'inboxes_promotion_status'))
+        {
+            Schema::table('doctor_inboxes', function (Blueprint $table) {
+                $table->dropColumn('inboxes_promotion_status');
+                $table->string('inboxes_category')->after('inboxes_id_reference');
+            });
+        }
     }
 
     /**
